Updated example
As discussed in #132 with @ailiop, I removed the installation example from Getting started and integrated it into the Install page. It differs from the previous example in a couple important respects. It does not use a version-specific subdirectory, and it produces no license language to confirm (probably because it uses `--exclude-subdir`).